Move "random" utils to own module.
